Moved the application into its own module
