+++ README.md
@@ -6,10 +6,10 @@ h2. TODO
 
 * Add field of view control
 * Fix box height
-* refactor the application into its own function
 * add Jest tests
 * add texture rendering
 * tidy up readme
 * add player collision
 * add strafe
 * fix line rendering error on middle line when starting into the app
+* Add key pressed handling
